+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 
 * Fix the behavior of building empty projects so that when no models can be
   found, the Gradle plugin warns instead of fails.
+  
+### Features
+
+* Scan for a smithy-cli buildScript dependency when building models if no explicit
+  smithy-cli or smithy-model dependency can be found.